Keep the channel core/runtime split mechanically sound before landing
This rebases the core/runtime extraction onto the latest dev branch and
keeps the refactor focused while fixing the concrete issues that belong
inside this slice: command metadata stays in the command layer,
serve-runtime teardown now preserves shutdown failures, the shared send
helper no longer mutates process-wide runtime env state, the
Future/Pin gates now match ChannelCommandFuture users, and the slow
redirect test no longer burns a fixed startup sleep.

Constraint: Keep the channel refactor behavior-preserving and compatible with the current public surface
Rejected: Change session_key participant/thread encoding in this slice | compatibility risk beyond the refactor boundary
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep command-family metadata out of runtime helpers and preserve shutdown failures when touching serve-runtime orchestration
Tested: cargo fmt --all -- --check
Tested: cargo check -p loongclaw-app --all-features
Tested: cargo clippy -p loongclaw-app --all-features -- -D warnings
Tested: cargo test -p loongclaw-app merge_runtime_result_preserves_both_errors --all-features
Tested: cargo test -p loongclaw-app outbound_http_client_returns_redirect_status_after_slow_start --all-features
Not-tested: Full workspace all-features test matrix locally; waiting on GitHub CI for the broad matrix
Related: #892
