+++ crates/app/src/channel/commands/mod.rs
@@ -4,4 +4,11 @@ mod serve;
 
 pub(super) use context::{ChannelCommandContext, ChannelResolvedRuntimeAccount};
 pub(super) use send::{ChannelSendCommandSpec, run_channel_send_command};
-pub(super) use serve::run_channel_serve_command_with_stop;
+#[cfg(any(
+    feature = "channel-telegram",
+    feature = "channel-feishu",
+    feature = "channel-matrix",
+    feature = "channel-wecom",
+    feature = "channel-whatsapp"
+))]
+pub(super) use serve::{ChannelServeCommandSpec, run_channel_serve_command_with_stop};
