+++ crates/app/src/channel/core/mod.rs
@@ -1,0 +1,5 @@
+// crates/app/src/channel/core/mod.rs
+
+pub(in crate::channel) mod http;
+pub(in crate::channel) mod types;
+pub(in crate::channel) mod webhook_auth;
