+++ crates/app/src/channel/runtime/mod.rs
@@ -1,0 +1,7 @@
+// crates/app/src/channel/runtime/mod.rs
+
+pub(in crate::channel) mod http;
+pub(in crate::channel) mod serve;
+pub(in crate::channel) mod state;
+pub(in crate::channel) mod turn_feedback;
+pub(in crate::channel) mod types;
