+++ crates/app/src/channel/runtime/serve.rs
@@ -9,10 +9,9 @@ use std::{
 
 use tokio::sync::Notify;
 
-use super::registry::ChannelCommandFamilyDescriptor;
-use super::runtime_state;
-use super::runtime_state::ChannelOperationRuntimeTracker;
-use super::types::ChannelPlatform;
+use super::super::types::ChannelPlatform;
+use super::state;
+use super::state::ChannelOperationRuntimeTracker;
 use crate::CliResult;
 
 #[cfg(any(
@@ -23,23 +22,11 @@ use crate::CliResult;
     feature = "channel-whatsapp"
 ))]
 #[derive(Debug, Clone, Copy)]
-pub(super) struct ChannelServeRuntimeSpec<'a> {
-    pub(super) platform: ChannelPlatform,
-    pub(super) operation_id: &'static str,
-    pub(super) account_id: &'a str,
-    pub(super) account_label: &'a str,
-}
-
-#[cfg(any(
-    feature = "channel-telegram",
-    feature = "channel-feishu",
-    feature = "channel-matrix",
-    feature = "channel-wecom",
-    feature = "channel-whatsapp"
-))]
-#[derive(Debug, Clone, Copy)]
-pub(super) struct ChannelServeCommandSpec {
-    pub(super) family: ChannelCommandFamilyDescriptor,
+pub(in crate::channel) struct ChannelServeRuntimeSpec<'a> {
+    pub(in crate::channel) platform: ChannelPlatform,
+    pub(in crate::channel) operation_id: &'static str,
+    pub(in crate::channel) account_id: &'a str,
+    pub(in crate::channel) account_label: &'a str,
 }
 
 #[derive(Debug, Clone)]
@@ -72,7 +59,7 @@ impl ChannelServeStopHandle {
         feature = "channel-wecom",
         feature = "channel-whatsapp"
     ))]
-    pub(super) async fn wait(&self) {
+    pub(in crate::channel) async fn wait(&self) {
         if self.is_requested() {
             return;
         }
@@ -92,7 +79,7 @@ impl ChannelServeStopHandle {
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-pub(super) fn channel_runtime_now_ms() -> u64 {
+pub(in crate::channel) fn channel_runtime_now_ms() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .map(|value| value.as_millis() as u64)
@@ -106,19 +93,19 @@ pub(super) fn channel_runtime_now_ms() -> u64 {
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-pub(super) fn ensure_channel_operation_runtime_slot_available_in_dir(
+pub(in crate::channel) fn ensure_channel_operation_runtime_slot_available_in_dir(
     runtime_dir: &std::path::Path,
     spec: ChannelServeRuntimeSpec<'_>,
 ) -> CliResult<()> {
     let now = channel_runtime_now_ms();
-    runtime_state::prune_inactive_channel_operation_runtime_files_for_account_from_dir(
+    state::prune_inactive_channel_operation_runtime_files_for_account_from_dir(
         runtime_dir,
         spec.platform,
         spec.operation_id,
         Some(spec.account_id),
         now,
     )?;
-    let Some(runtime) = runtime_state::load_channel_operation_runtime_for_account_from_dir(
+    let Some(runtime) = state::load_channel_operation_runtime_for_account_from_dir(
         runtime_dir,
         spec.platform,
         spec.operation_id,
@@ -152,7 +139,7 @@ pub(super) fn ensure_channel_operation_runtime_slot_available_in_dir(
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-pub(super) async fn with_channel_serve_runtime<T, F, Fut>(
+pub(in crate::channel) async fn with_channel_serve_runtime<T, F, Fut>(
     spec: ChannelServeRuntimeSpec<'_>,
     run: F,
 ) -> CliResult<T>
@@ -161,7 +148,7 @@ where
     Fut: Future<Output = CliResult<T>>,
 {
     ensure_channel_operation_runtime_slot_available_in_dir(
-        runtime_state::default_channel_runtime_state_dir().as_path(),
+        state::default_channel_runtime_state_dir().as_path(),
         spec,
     )?;
     let runtime = Arc::new(
@@ -175,13 +162,7 @@ where
     );
     let result = run(runtime.clone()).await;
     let shutdown_result = runtime.shutdown().await;
-    match result {
-        Err(error) => Err(error),
-        Ok(value) => {
-            shutdown_result?;
-            Ok(value)
-        }
-    }
+    merge_runtime_result(result, shutdown_result)
 }
 
 #[cfg(any(
@@ -191,7 +172,7 @@ where
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-pub(super) async fn with_channel_serve_runtime_with_stop<F, Fut>(
+pub(in crate::channel) async fn with_channel_serve_runtime_with_stop<F, Fut>(
     spec: ChannelServeRuntimeSpec<'_>,
     stop: ChannelServeStopHandle,
     run: F,
@@ -204,7 +185,7 @@ where
 }
 
 #[cfg(test)]
-pub(super) async fn with_channel_serve_runtime_with_stop_in_dir<F, Fut>(
+pub(in crate::channel) async fn with_channel_serve_runtime_with_stop_in_dir<F, Fut>(
     runtime_dir: &std::path::Path,
     process_id: u32,
     spec: ChannelServeRuntimeSpec<'_>,
@@ -222,7 +203,7 @@ where
 }
 
 #[cfg(test)]
-pub(super) async fn with_channel_serve_runtime_in_dir<T, F, Fut>(
+pub(in crate::channel) async fn with_channel_serve_runtime_in_dir<T, F, Fut>(
     runtime_dir: &std::path::Path,
     process_id: u32,
     spec: ChannelServeRuntimeSpec<'_>,
@@ -234,7 +215,7 @@ where
 {
     ensure_channel_operation_runtime_slot_available_in_dir(runtime_dir, spec)?;
     let runtime = Arc::new(
-        runtime_state::start_channel_operation_runtime_tracker_for_test(
+        state::start_channel_operation_runtime_tracker_for_test(
             runtime_dir,
             spec.platform,
             spec.operation_id,
@@ -246,11 +227,54 @@ where
     );
     let result = run(runtime.clone()).await;
     let shutdown_result = runtime.shutdown().await;
-    match result {
-        Err(error) => Err(error),
-        Ok(value) => {
-            shutdown_result?;
-            Ok(value)
-        }
+    merge_runtime_result(result, shutdown_result)
+}
+
+fn merge_runtime_result<T>(result: CliResult<T>, shutdown_result: CliResult<()>) -> CliResult<T> {
+    match (result, shutdown_result) {
+        (Ok(value), Ok(())) => Ok(value),
+        (Ok(_), Err(shutdown_error)) => Err(shutdown_error),
+        (Err(run_error), Ok(())) => Err(run_error),
+        (Err(run_error), Err(shutdown_error)) => Err(format!(
+            "{run_error}; additionally failed to shut down channel runtime cleanly: {shutdown_error}"
+        )),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::merge_runtime_result;
+
+    #[test]
+    fn merge_runtime_result_returns_value_when_both_steps_succeed() {
+        let result = merge_runtime_result::<u8>(Ok(7), Ok(()));
+
+        assert_eq!(result, Ok(7));
+    }
+
+    #[test]
+    fn merge_runtime_result_returns_shutdown_error_after_successful_run() {
+        let result = merge_runtime_result::<u8>(Ok(7), Err("shutdown failed".to_owned()));
+
+        assert_eq!(result, Err("shutdown failed".to_owned()));
+    }
+
+    #[test]
+    fn merge_runtime_result_returns_run_error_when_shutdown_succeeds() {
+        let result = merge_runtime_result::<u8>(Err("run failed".to_owned()), Ok(()));
+
+        assert_eq!(result, Err("run failed".to_owned()));
+    }
+
+    #[test]
+    fn merge_runtime_result_preserves_both_errors() {
+        let result = merge_runtime_result::<u8>(
+            Err("run failed".to_owned()),
+            Err("shutdown failed".to_owned()),
+        );
+        let error = result.expect_err("both failures should produce an error");
+
+        assert!(error.contains("run failed"));
+        assert!(error.contains("shutdown failed"));
     }
 }
