+++ crates/app/src/channel/whatsapp/mod.rs
@@ -7,14 +7,14 @@ use axum::{Router, routing::get};
 use serde_json::{Value, json};
 
 use super::dispatch::{
-    ChannelCommandContext, ChannelResolvedRuntimeAccount, build_whatsapp_command_context,
-    run_channel_serve_command_with_stop,
+    ChannelCommandContext, ChannelResolvedRuntimeAccount, ChannelServeCommandSpec,
+    build_whatsapp_command_context, run_channel_serve_command_with_stop,
 };
-use super::serve_runtime::{ChannelServeCommandSpec, ChannelServeStopHandle};
+use super::runtime::serve::ChannelServeStopHandle;
 use super::{ChannelOutboundTargetKind, WHATSAPP_COMMAND_FAMILY_DESCRIPTOR};
 use super::{
     http::{ChannelOutboundHttpPolicy, build_outbound_http_client, validate_outbound_http_target},
-    runtime_state::ChannelOperationRuntimeTracker,
+    runtime::state::ChannelOperationRuntimeTracker,
 };
 use crate::config::{ChannelDefaultAccountSelectionSource, LoongClawConfig};
 use crate::{CliResult, KernelContext, config::ResolvedWhatsappChannelConfig};
